Move data conversion to G6Renderer
