+++ src/component/g6/node/gateway-nodes.ts
