+++ src/component/g6/node/text-annotation-nodes.ts
