+++ src/component/g6/shape/IconPainter.ts
