+++ src/component/g6/shape/index.ts
