+++ src/component/g6/shape/render-types.ts
