+++ src/component/g6/style-helper.ts
@@ -13,9 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-import { mxCell } from 'mxgraph'; // for types
-
 import { Item } from '@antv/g6/lib/types';
 
 /**
