+++ test/unit/component/mxgraph/shape/render/BpmnCanvas.test.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { IconStyleConfiguration, ShapeConfiguration, Size, computeScaledIconSize } from '../../../../../../src/component/g6/shape/render';
+import { IconStyleConfiguration, ShapeConfiguration, Size, computeScaledIconSize } from '../../../../../../src/component/g6/shape';
 
 describe('compute scaled icon size', () => {
   function newIconStyleConfiguration(strokeWidth = 0): IconStyleConfiguration {
